+++ Wolf3D/src/vertical.c
@@ -12,7 +12,7 @@
 
 #include "wolf3d.h"
 
-t_point *getVertical(t_tool *t, float rayAngle)
+t_point *getVertical(t_tool *t, double rayAngle)
 {
   t_point *inc;
   t_point *curPoint;
@@ -28,7 +28,7 @@ t_point *getVertical(t_tool *t, float rayAngle)
   return curPoint;
 }
 
-t_point *getFirstPointVertical(t_tool *t, float rayAngle)
+t_point *getFirstPointVertical(t_tool *t, double rayAngle)
 {
   t_point *point;
 
@@ -49,7 +49,7 @@ t_point *getFirstPointVertical(t_tool *t, float rayAngle)
 t_point *getNextPointVertical(t_tool *t, t_point *curPoint, t_point *inc)
 {
   t_point *nextPoint;
-  float dist;
+  double dist;
 
   nextPoint = malloc(sizeof(*nextPoint));
   nextPoint->x = curPoint->x + inc->x;
